requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
pip==22.0.4
pydantic==1.9.0
PyJWT==2.3.0
pytz==2022.1
setuptools==60.9.3
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
wheel==0.37.1
gunicorn==20.1.0
